implemented view method for decoratedPatient
