+++ src/main/java/com/laba/controllers/EntityController.java
@@ -1,4 +1,4 @@
-package com.laba.controllers.controller;
+package com.laba.controllers;
 
 import com.laba.interfaces.controllers.IEntityController;
 import com.laba.utils.json.JacksonUtil;
@@ -6,8 +6,8 @@ import com.laba.view.EntityView;
 
 public abstract class EntityController<T, D extends EntityView<T>> implements IEntityController<T> {
 
-    private T model;
-    private D view;
+    protected T model;
+    protected D view;
 
 
     public EntityController(T model, D view) {
