+++ src/main/java/com/laba/view/DecoratedPatientView.java
@@ -4,4 +4,9 @@ import com.laba.models.DecoratedPatient;
 
 public class DecoratedPatientView extends EntityView<DecoratedPatient> {
 
+    public void simplePatientMedicalRecordsSysOutView(String data) {
+        System.out.println("Simplified Medical Record:");
+        System.out.println(data);
+    }
+
 }
